requirements.txt
fastapi
uvicorn
streamlit
python-multipart
requests
python-dotenv
langgraph
mcp
pymupdf
httpx
mistralai